Create setwd
MImics bash  parameters which allow you to include the current directory path as part of the prompt.  I haven't decided whether I prefer to overload "setwd" or to save this as something like "petwd" to give me the option of resetting or not resetting the prompt.  Feel free to take a copy and mod as you wish.