Create splatnd
A sneaky operator which lets you execute various commands sans arguments (or even the parentheses).   This is included in the CRAN package "cgwtools"  ;  I posted it here for others to play with.  The intent is for everyone to add whatever commands they desire to the "switch" command inside this code.